% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidierplots.R
\name{scale_colour_continuous}
\alias{scale_colour_continuous}
\title{scale_colour_continuous}
\usage{
scale_colour_continuous(...)
}
\description{
TidierPlots wrapper for ggplot2::scale_colour_continuous.
}
